# Conflicts:
#	service/src/main/java/com/nicico/training/service/EvaluationAnalysisService.java

+++ service/src/main/java/com/nicico/training/service/ClassStudentService.java
@@ -11,6 +11,7 @@ import com.nicico.training.iservice.ITclassService;
 import com.nicico.training.model.ClassStudent;
 import com.nicico.training.model.Student;
 import com.nicico.training.model.Tclass;
+import com.nicico.training.repository.AttendanceDAO;
 import com.nicico.training.repository.ClassStudentDAO;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -25,14 +26,14 @@ import java.util.function.Function;
 public class ClassStudentService implements IClassStudentService {
 
     private final ClassStudentDAO classStudentDAO;
+    private final AttendanceDAO attendanceDAO;
     private final ITclassService tclassService;
     private final StudentService studentService;
 //    private final IPersonnelService personnelService;
     private final IPersonnelRegisteredService personnelRegisteredService;
     private final ModelMapper mapper;
     private final IEvaluationAnalysisService evaluationAnalysisService;
-    private final ViewActivePersonnelService activePersonnelService;
-
+    private final PersonnelService personnelService;
 
 
     @Transactional(readOnly = true)
@@ -55,7 +56,7 @@ public class ClassStudentService implements IClassStudentService {
         Tclass tclass = tclassService.getTClass(classId);
 
         for (ClassStudentDTO.Create c : request) {
-            List<Student> list = studentService.getStudentByPostCodeAndPersonnelNoAndDepartmentCodeAndFirstNameAndLastName(c.getPostCode(), c.getPersonnelNo(), c.getDepartmentCode(), c.getFirstName(), c.getLastName());
+            List<Student> list = studentService.getStudentByPostIdAndPersonnelNoAndDepartmentIdAndFirstNameAndLastNameOrderByIdDesc(c.getPostId(), c.getPersonnelNo(), c.getDepartmentId(), c.getFirstName(), c.getLastName());
             Student student = null;
             int size = list.size();
             for (int i=0; i<size; i++) {
@@ -70,10 +71,10 @@ public class ClassStudentService implements IClassStudentService {
             if(student == null) {
                 student = new Student();
                 if (c.getRegisterTypeId() == 1) {
-                    mapper.map(activePersonnelService.getByPersonnelCode(c.getPersonnelNo()), student);
+                    mapper.map(personnelService.getByPersonnelCodeAndNationalCode(c.getNationalCode(),  c.getPersonnelNo()), student);
                     student.setDeleted(null);
                 } else if (c.getRegisterTypeId() == 2) {
-                    mapper.map(personnelRegisteredService.getByPersonnelCode(c.getPersonnelNo()), student);
+                    mapper.map(personnelRegisteredService.getByPersonnelCodeAndNationalCode(c.getNationalCode(), c.getPersonnelNo()), student);
                 }
 
                 /*SearchDTO.SearchRq searchRq = new SearchDTO.SearchRq();
@@ -123,8 +124,16 @@ public class ClassStudentService implements IClassStudentService {
 
     @Transactional
     @Override
-    public void delete(Long id) {
-        classStudentDAO.deleteById(id);
+    public String delete(Long id) {
+
+        ClassStudent classSession = classStudentDAO.getClassStudentById(id);
+        if (attendanceDAO.checkAttendanceByStudentIdAndClassId(classSession.getStudentId(), classSession.getTclassId()) > 0) {
+            return "فراگير «<b>" + classSession.getStudent().getFirstName() + " " + classSession.getStudent().getLastName() + "</b>» بدلیل داشتن حضور و غیاب قابل حذف نیست.";
+        } else {
+            classStudentDAO.deleteById(id);
+            return "";
+        }
+
     }
 
     @Transactional
